Make gridder into a package
Separated the code in fatiando.gridder into several modules that live
inside a gridder package. All functions are imported into the
fatiando.gridder namespace so the user doesn't see any changes. This
will make the code easier to manage going forward.

A function that had no home is load_surfer which should be moved into
the future fatiando.datasets package. For now, I left it in the
__init__.py  file.

Moved the tests regarding gridder into the gridder/tests folder. This
makes it easier to run all tests+doctests from gridder without running
the whole test suite with:

    py.test --doctest-modules fatiando/gridder

The only problem still is the documentation. The links to all functions
are broken and automodule doesn't render their docstrings in the module
documentation. This should not be a problem actually because we'll move
all the API docs into a single list of functions and classes. Will see
what I can do about this for now.